Ignore parameters when validating Content-Type MIME type
The Swagger specification says that consumes and produces are "A list
of MIME types...". The Content-Type header of a HTTP request consists
of a MIME type *and* optional parameters, e.g.

    text/html; charset=utf-8

validation of request and response content type was failing when
parameters were present as the entire value of the Content-Type header
was compared to the allowed MIME types in consumes/produces.

This changes the validation of content types to only compare the MIME
type in the header to the allowed MIME types defined in consumes and
produces as appropriate for requests/responses.