Do not validate Content-Type if it is missing
This is the behaviour flex had prior to d91debaa, in that commit the
Content-Type validation logic was changed to assume the Content-Type
of a request was application/octet-stream if it was not otherwise
given. This is not wrong per se, but as the Content-Type header is
not given with requests with no body (e.g. a GET request) this meant
that these requests were considered to have the an invalid
Content-Type when they really, correctly had provided None. This
restores the previous behaviour where, in all cases, Content-Type
was would be ignored if not provided.