Validate date/times are RFC3339 compliant
The Swagger spec states that date/times must be in the format described
by RFC3339, this a strict subset of the ISO8601 format, essentially
requiring that date/times are of the form:

    2017-01-02T12:34:56Z
    or
    2017-01-02T12:34:56+00:00

(with microseconds allowed but optional).

Previously flex validated date/times against ISO8601, this is more
permissive than RFC3339 and so some date/time strings were considered
valid as they were valid ISO8601 date/times but were not valid
against RFC3339.

This changes date/time validation to check against the format specified
in RFC3339.