Avoid AttributeError on request/response data when Content-Type missing
HTTP POST/PUT requests SHOULD include a Content-Type but they don't
have to, if flex tried to validate a request without a Content-Type
an AttributeError would be raised as the data property of Request
and Response checks the content_type with .startswith to see if
it's JSON (and so should be parsed before further processing). This
exception happened before any validation on Content-Type so if you
have a schema with an endpoint that consumes application/json but
you validate a request without any Content-Type you ended up with
this exception instead of some validation error.

This puts a simple guard before the .startswith call preventing the
AttributeError.